requirements-colab.txt
git+https://github.com/facebookresearch/segment-anything.git
groundingdino-py
transformers==4.41.2
diffusers==0.29.2
peft==0.11.1
accelerate
simple-lama-inpainting
pvlib
scikit-image
scipy
opencv-python
matplotlib
numpy
torch
torchvision
